+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireViewModel.kt
@@ -27,6 +27,7 @@ import com.google.android.fhir.datacapture.mapping.ResourceMapper
 import com.google.android.fhir.datacapture.targetStructureMap
 import com.google.android.fhir.logicalId
 import dagger.hilt.android.lifecycle.HiltViewModel
+import java.util.Calendar
 import java.util.Date
 import java.util.UUID
 import javax.inject.Inject
@@ -54,10 +55,12 @@ import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
 import org.smartregister.fhircore.engine.util.USER_INFO_SHARED_PREFERENCE_KEY
 import org.smartregister.fhircore.engine.util.extension.decodeJson
 import org.smartregister.fhircore.engine.util.extension.deleteRelatedResources
+import org.smartregister.fhircore.engine.util.extension.find
 import org.smartregister.fhircore.engine.util.extension.isIn
 import org.smartregister.fhircore.engine.util.extension.prepareQuestionsForReadingOrEditing
 import org.smartregister.fhircore.engine.util.extension.retainMetadata
 import org.smartregister.fhircore.engine.util.helper.TransformSupportServices
+import timber.log.Timber
 
 @HiltViewModel
 open class QuestionnaireViewModel
@@ -145,6 +148,13 @@ constructor(
 
               if (resource is Patient) resource.managingOrganization = organizationRef
               else if (resource is Group) resource.managingEntity = organizationRef
+              // TODO: calculate birthDate from AgeInput as calculation isn't supported by sdk
+              // https://github.com/google/android-fhir/issues/803
+              if (resource is Patient) {
+                getAgeInput(questionnaireResponse)?.let {
+                  resource.birthDate = calculateDobFromAge(it)
+                }
+              }
             }
           }
 
@@ -270,7 +280,9 @@ constructor(
                 system = QuestionnaireActivity.WHO_IDENTIFIER_SYSTEM
               }
             )
+          Timber.e(FhirContext.forR4().newJsonParser().encodeResourceToString(this))
         }
+
         resourcesList.add(this)
       }
       loadRelatedPerson(patientId)?.forEach { resourcesList.add(it) }
@@ -284,5 +296,24 @@ constructor(
     intent: Intent
   ): QuestionnaireResponse {
     return ResourceMapper.populate(questionnaire, *getPopulationResources(intent))
+  }
+
+  fun getAgeInput(questionnaireResponse: QuestionnaireResponse): Int? {
+    return questionnaireResponse
+      .find(QuestionnaireActivity.QUESTIONNAIRE_AGE)
+      ?.answer
+      ?.firstOrNull()
+      ?.valueDecimalType
+      ?.value
+      ?.toInt()
+  }
+
+  fun calculateDobFromAge(age: Int): Date {
+    val cal: Calendar = Calendar.getInstance()
+    // Subtract #age years from the calendar
+    cal.add(Calendar.YEAR, -age)
+    cal.set(Calendar.DAY_OF_YEAR, 1)
+    cal.set(Calendar.MONTH, 1)
+    return cal.time
   }
 }