+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientDetailScreenTest.kt
@@ -43,7 +43,6 @@ import javax.inject.Inject
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.junit.Assert.assertEquals
 import org.junit.Before
-import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireConfig
@@ -95,6 +94,7 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testToolbarComponents() {
+    initMocks()
     composeRule
       .onNodeWithTag(TOOLBAR_TITLE)
       .assertTextEquals(application.getString(R.string.back_to_clients))
@@ -110,6 +110,7 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testToolbarMenuButtonShouldToggleMenuItemsList() {
+    initMocks()
     composeRule.onNodeWithTag(TOOLBAR_MENU).assertDoesNotExist()
     composeRule.onNodeWithTag(TOOLBAR_MENU_BUTTON).performClick()
 
@@ -126,6 +127,7 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testToolbarTestResultsMenuItemShouldCallMenuItemClickListener() {
+    initMocks()
     composeRule.onNodeWithTag(TOOLBAR_MENU_BUTTON).performClick()
     composeRule.onNodeWithTag(TOOLBAR_MENU).onChildAt(0).performClick()
     verify { questPatientDetailViewModel.onMenuItemClickListener(R.string.test_results) }
@@ -140,12 +142,14 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testToolbarBackPressedButtonShouldCallBackPressedClickListener() {
+    initMocks()
     composeRule.onNodeWithTag(TOOLBAR_BACK_ARROW).performClick()
     verify { questPatientDetailViewModel.onBackPressed(true) }
   }
 
   @Test
   fun testPatientDetailsCardShouldHaveCorrectData() {
+    initMocks()
     composeRule
       .onNodeWithTag(PATIENT_NAME)
       .assertExists()
@@ -155,6 +159,7 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testFormsListShouldDisplayAllFormWithCorrectData() {
+    initMocks()
     composeRule.onAllNodesWithTag(FORM_ITEM).assertCountEquals(2)
     composeRule.onAllNodesWithTag(FORM_ITEM)[0].assertTextEquals("SAMPLE ORDER RESULT")
     composeRule.onAllNodesWithTag(FORM_ITEM)[1].assertTextEquals("SAMPLE TEST RESULT")
@@ -162,6 +167,7 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
 
   @Test
   fun testFormsListItemClickShouldCallFormItemClickListener() {
+    initMocks()
     val formItemClicks = mutableListOf<QuestionnaireConfig>()
     every { questPatientDetailViewModel.onFormItemClickListener(any()) } answers
       {
@@ -179,8 +185,9 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
   }
 
   @Test
-  @Ignore("Fix assertions once questionnaire listing bug is resolved")
   fun testResultsListShouldHaveAllResultsWithCorrectData() {
+    initMocks()
+
     // response heading should be exist and must be displayed
     composeRule.onNodeWithText("RESPONSES (2)").assertExists().assertIsDisplayed()
 
@@ -195,8 +202,8 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
   }
 
   @Test
-  @Ignore("Fix assertions once questionnaire listing bug is resolved")
   fun testResultsListItemShouldCallResultItemListener() {
+    initMocks()
     val resultItemClicks = mutableListOf<QuestionnaireResponse>()
     every { questPatientDetailViewModel.onTestResultItemClickListener(any()) } answers
       {
@@ -214,16 +221,48 @@ class QuestPatientDetailScreenTest : RobolectricTest() {
   }
 
   @Test
-  @Ignore("Fix assertions once questionnaire listing bug is resolved")
   fun testQuestPatientDetailScreenShouldHandleMissingData() {
+    initEmptyMocks()
     composeRule
       .onNodeWithTag(TOOLBAR_TITLE)
       .assertTextEquals(application.getString(R.string.back_to_clients))
     composeRule.onNodeWithTag(TOOLBAR_BACK_ARROW).assertHasClickAction()
-    composeRule.onNodeWithTag(PATIENT_NAME).assertExists().assertTextEquals("")
+    composeRule.onNodeWithTag(PATIENT_NAME).assertExists().assertTextEquals(", , ")
     composeRule.onNodeWithTag(FORM_CONTAINER_ITEM).assert(hasAnyChild(hasText("Loading forms ...")))
     composeRule
       .onNodeWithTag(RESULT_CONTAINER_ITEM)
       .assert(hasAnyChild(hasText("Loading responses ...")))
+  }
+
+  private fun initMocks() {
+    Faker.initPatientRepositoryMocks(patientRepository)
+    questPatientDetailViewModel =
+      spyk(
+        QuestPatientDetailViewModel(
+          patientRepository = patientRepository,
+          patientItemMapper = patientItemMapper
+        )
+      )
+    // Simulate retrieval of data from repository
+    questPatientDetailViewModel.run {
+      getDemographics(patientId)
+      getAllResults(patientId)
+      getAllForms(application)
+    }
+
+    composeRule.setContent { QuestPatientDetailScreen(questPatientDetailViewModel) }
+  }
+
+  private fun initEmptyMocks() {
+    Faker.initPatientRepositoryEmptyMocks(patientRepository)
+
+    questPatientDetailViewModel =
+      spyk(
+        QuestPatientDetailViewModel(
+          patientRepository = patientRepository,
+          patientItemMapper = patientItemMapper
+        )
+      )
+    composeRule.setContent { QuestPatientDetailScreen(questPatientDetailViewModel) }
   }
 }